+++ API_GATEWAY/src/schemas/createSession.schema.ts
